rename kafka.protocol.offset -> kafka.protocol.list_offsets
